Move http to sender, move Rollbar to rollbar
