+++ pom.xml
@@ -16,5 +16,6 @@
     <module>rollbar-sender</module>
     <module>rollbar-utilities</module>
     <module>rollbar-testing</module>
+    <module>rollbar</module>
   </modules>
 </project>
