+++ rollbar-sender/src/main/java/com/rollbar/sender/RollbarResponseHandler.java
@@ -1,4 +1,4 @@
-package com.rollbar.http;
+package com.rollbar.sender;
 
 public interface RollbarResponseHandler {
     void handleResponse(RollbarResponse response);
