+++ rollbar-sender/src/main/java/com/rollbar/sender/Sender.java
@@ -1,4 +1,4 @@
-package com.rollbar.http;
+package com.rollbar.sender;
 
 import com.rollbar.payload.Payload;
