+++ rollbar-sender/src/test/java/com/rollbar/sender/InvalidResponseCodeExceptionTest.java
@@ -1,4 +1,4 @@
-package com.rollbar.http;
+package com.rollbar.sender;
 
 import org.junit.Test;
