+++ rollbar/pom.xml
@@ -8,16 +8,24 @@
         <version>0.0.1-alpha</version>
         <relativePath>../parent-pom.xml</relativePath>
     </parent>
-
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>rollbar-sender</artifactId>
-    <version>0.0.1-alpha</version>
+    <artifactId>rollbar</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>com.rollbar</groupId>
             <artifactId>rollbar-payload</artifactId>
+            <version>0.0.1-alpha</version>
+        </dependency>
+        <dependency>
+            <groupId>com.rollbar</groupId>
+            <artifactId>rollbar-sender</artifactId>
+            <version>0.0.1-alpha</version>
+        </dependency>
+        <dependency>
+            <groupId>com.rollbar</groupId>
+            <artifactId>rollbar-utilities</artifactId>
             <version>0.0.1-alpha</version>
         </dependency>
     </dependencies>
