+++ rollbar/src/main/java/com/rollbar/PayloadFilter.java
