+++ rollbar/src/main/java/com/rollbar/PayloadTransform.java
